+++ server/src/main/java/com/hobbyt/domain/user/dto/request/EmailRequest.java
@@ -1,6 +1,7 @@
-package com.hobbyt.domain.user.dto;
+package com.hobbyt.domain.user.dto.request;
 
 import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -11,5 +12,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class EmailRequest {
 	@Email
+	@NotBlank
 	private String email;
 }
