+++ server/src/main/java/com/hobbyt/domain/user/service/AuthService.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.hobbyt.domain.user.dto.EmailRequest;
+import com.hobbyt.domain.user.dto.request.EmailRequest;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
